refactor(ListItem) move from HOCs to hooks
