--- packages/core/src/components/list-item/ListItem.ts
@@ -1,18 +1,0 @@
-/**
- * Copyright (c) Flavio Silva https://flsilva.com
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import * as React from 'react';
-
-// tslint:disable-next-line:max-line-length
-import { withResponsivenessProps } from '../../responsiveness/withResponsivenessProps';
-import { ListItemPropsOptional } from './ListItemProps';
-import { SimpleListItem } from './SimpleListItem';
-import { withDefaultListItemProps } from './withDefaultListItemProps';
-
-export const ListItem: React.ComponentType<
-  ListItemPropsOptional
-> = withResponsivenessProps(withDefaultListItemProps(SimpleListItem));
